refactor(components): use css var to set menu global style (#3539)
* refactor(components): use css var to set menu global style

* fix(components): remove useless value & change camelCase & test file

* test: comment code that can not test final style

* test: fix unused-vars lint

* feat: add computed for useMenuCssVar

* test(components): revert test file

* fix(components): use computed value
