+++ DEV_FAQ.md
@@ -1,0 +1,19 @@
+# DEV FAQ
+
+Here are the problems that are easy to encounter in development.
+
+## If you encounter dependency related issues
+
+```bash
+yarn bootstrap
+```
+
+## Test suite failed to run (jest)
+
+Can not run `yarn test xxx`
+
+```bash
+rm -rf dist
+```
+
+Try again.
