+++ packages/theme-chalk/src/common/var.scss
@@ -896,10 +896,13 @@ $--slider: map.merge(
 $--menu: () !default;
 $--menu: map.merge(
   (
-    'item-font-size': var(--el-font-size-base),
-    'item-font-color': var(--el-text-color-primary),
-    'item-hover-fill': var(--el-color-primary-light-9),
+    'active-color': var(--el-color-primary),
+    'text-color': var(--el-text-color-primary),
+    'hover-text-color': var(--el-text-color-primary),
     'background-color': var(--el-color-white),
+    'hover-background-color': var(--el-color-primary-light-9),
+    'item-font-size': var(--el-font-size-base),
+    'item-hover-fill': var(--el-color-primary-light-9),
     'border-color': #e6e6e6,
   ),
   $--menu
