+++ packages/theme-chalk/src/menu.scss
@@ -8,7 +8,7 @@
   height: 56px;
   line-height: 56px;
   font-size: var(--el-menu-item-font-size);
-  color: var(--el-menu-item-font-color);
+  color: var(--el-menu-text-color);
   padding: 0 20px;
   list-style: none;
   cursor: pointer;
@@ -24,7 +24,7 @@
   }
 
   i {
-    color: var(--el-text-color-secondary);
+    color: inherit;
   }
 
   &:hover,
@@ -33,7 +33,7 @@
   }
 
   &:hover {
-    background-color: var(--el-menu-item-hover-fill);
+    background-color: var(--el-menu-hover-background-color);
   }
 
   @include when(disabled) {
@@ -68,7 +68,7 @@
       height: 60px;
       margin: 0;
       border-bottom: 2px solid transparent;
-      color: var(--el-text-color-secondary);
+      color: var(--el-menu-text-color);
 
       a,
       a:hover {
@@ -88,14 +88,14 @@
 
       &:hover {
         .#{$namespace}-sub-menu__title {
-          color: var(--el-text-color-primary);
+          color: var(--el-menu-hover-text-color);
         }
       }
 
       &.is-active {
         .#{$namespace}-sub-menu__title {
-          border-bottom: 2px solid var(--el-color-primary);
-          color: var(--el-text-color-primary);
+          border-bottom: 2px solid var(--el-menu-active-color);
+          color: var(--el-menu-active-color);
         }
       }
 
@@ -103,7 +103,7 @@
         height: 60px;
         line-height: 60px;
         border-bottom: 2px solid transparent;
-        color: var(--el-text-color-secondary);
+        color: var(--el-menu-text-color);
 
         &:hover {
           background-color: #fff;
@@ -120,27 +120,28 @@
     & .#{$namespace}-menu {
       & .#{$namespace}-menu-item,
       & .#{$namespace}-sub-menu__title {
-        background-color: $--color-white;
+        background-color: var(--el-menu-background-color);
         display: flex;
         align-items: center;
         height: 36px;
         padding: 0 10px;
-        color: var(--el-text-color-secondary);
+        color: var(--el-menu-text-color);
       }
 
       & .#{$namespace}-menu-item.is-active,
       & .#{$namespace}-sub-menu.is-active > .#{$namespace}-sub-menu__title {
-        color: var(--el-text-color-primary);
+        color: var(--el-menu-active-color);
       }
     }
     & .#{$namespace}-menu-item:not(.is-disabled):hover,
     & .#{$namespace}-menu-item:not(.is-disabled):focus {
       outline: none;
-      color: var(--el-text-color-primary);
+      color: var(--el-menu-hover-text-color);
+      background-color: var(--el-menu-hover-background-color);
     }
     & > .#{$namespace}-menu-item.is-active {
-      border-bottom: 2px solid var(--el-color-primary);
-      color: var(--el-text-color-primary);
+      border-bottom: 2px solid var(--el-menu-active-color);
+      color: var(--el-menu-active-color) !important;
     }
   }
   @include m(collapse) {
@@ -197,7 +198,7 @@
       }
 
       &.is-active i {
-        color: var(--el-color-primary);
+        color: inherit;
       }
     }
   }
@@ -222,7 +223,7 @@
     vertical-align: middle;
   }
   @include when(active) {
-    color: var(--el-color-primary);
+    color: var(--el-menu-active-color);
     i {
       color: inherit;
     }
@@ -238,7 +239,7 @@
     @include menu-item;
 
     &:hover {
-      background-color: var(--el-menu-item-hover-fill);
+      background-color: var(--el-menu-hover-background-color);
     }
   }
   & .#{$namespace}-menu {
@@ -268,7 +269,7 @@
   }
   @include when(active) {
     .#{$namespace}-sub-menu__title {
-      border-bottom-color: var(--el-color-primary);
+      border-bottom-color: var(--el-menu-active-color);
     }
   }
   @include when(opened) {
